add: cron

+++ server/api.server.js
@@ -13,6 +13,7 @@
 
 const {fetch} = require('react-fetch');
 // const {fetch} = require('node-fetch');
+const cron = require('node-cron');
 
 
 const register = require('react-server-dom-webpack/node-register');
@@ -307,6 +308,12 @@ app.delete(
     '/users/:id',
     handleErrors(async function (req, res) {
         // TODO: 認証
+        if (await auth(req.body.user_id, req.body.token) === false) {
+            sendResponse(req, res, null);
+            console.log(`------------ bookmarks 2 ------------`);
+
+            return;
+        }
 
         await pool.query('delete from users where id = $1', [req.params.id]);
         // await unlink(path.resolve(NOTES_PATH, `${req.params.id}.md`));
@@ -318,6 +325,12 @@ app.delete(
     '/bookmarks/:id',
     handleErrors(async function (req, res) {
         // TODO: 認証
+        if (await auth(req.body.user_id, req.body.token) === false) {
+            sendResponse(req, res, null);
+            console.log(`------------ bookmarks 2 ------------`);
+
+            return;
+        }
 
         console.log(`/bookmarks/:id ${req.params.id}`);
 
@@ -398,29 +411,28 @@ const getKey = () => {
         // https://console.cloud.google.com/apis/api/youtube.googleapis.com/credentials?project=
         // https://console.developers.google.com/apis/api/youtube.googleapis.com/overview?project=
 
-        // "AIzaSyCC-FYd9K-VhVZVzGOiJ_ltLPwck_1bkMc",
-        // "AIzaSyAtK1DbsKVmQPl8DDpyVe_7J_CLdEdEzps",
-
+        "AIzaSyCC-FYd9K-VhVZVzGOiJ_ltLPwck_1bkMc",
+        "AIzaSyAtK1DbsKVmQPl8DDpyVe_7J_CLdEdEzps",
         "AIzaSyCPTFcXn0V-0fEefMIAGrwUBg2o0urdU3E",
     ];
 
     return keys[Math.floor(Math.random() * keys.length)];
 }
+
 app.get(
     '/cron',
     handleErrors(async function (req, res) {
-
         getYouTubeData();
-
         res.json({
             result: "true",
         });
     })
 );
+
 function getYouTubeData() {
     const search = process.env.SEARCH;
     const key = getKey();
-    const endPointYouTube = `https://www.googleapis.com/youtube/v3/search?key=${key}&part=snippet&type=video&eventType=live&&maxResults=5&order=date&q=${search}`;
+    const endPointYouTube = `https://www.googleapis.com/youtube/v3/search?key=${key}&part=snippet&type=video&eventType=live&&maxResults=100&order=date&q=${search}`;
 
     console.log(`YouTube: ${endPointYouTube}`);
 
@@ -477,3 +489,12 @@ function getYouTubeData() {
         }
     });
 }
+
+cron.schedule('* * * * *', () => {
+    try{
+        getYouTubeData();
+    }
+    catch(e){
+        console.log(e.message);
+    }
+});
